Fixed bug in active item of widgets.
The "white-space: nowrap" makes the autocomplete widget fail when it has data with large labels that use two lines. It makes the label switch really fast between nowrap and default making it unusable. It still needs the wrap to make the full text viewable and is the default behavior in jQuery UI.